renamed files and changed resolver interface param
removed unnecessary underscore in file names

changed handling of resolver-specific parameters from function
parameters to global parameter dictionary
